requirements.txt
pandas==2.1.4
numpy==1.26.2
requests==2.31.0
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0
scikit-learn==1.3.2
scipy==1.11.4
pyspellchecker==0.8.1
fuzzywuzzy==0.18.0
python-Levenshtein==0.23.0
python-dotenv==1.0.0
colorama==0.4.6
rich==13.7.0
tqdm==4.66.1
tabulate==0.9.0
Flask==3.0.0
Flask-Cors==4.0.0
gunicorn==21.2.0
